% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{[.bootstrap}
\title{light bootstrap constructor}
\usage{
bootstrap(data, times = 50L)

\method{[}{bootstrap}(x, i)
}
\arguments{
\item{d}{a data.frame, tibble or matrix}
}
\description{
creates resampling indices but does not make entire copies of the dataset

simplifies accessing a bootstrap resample
}
\details{
use `x[i]` to access the i-th resample
}
