% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_map}
\alias{bootstrap_map}
\title{Apply a function to each entrie of bootstrap}
\usage{
bootstrap_map(x, .f, times = 50L)
}
\arguments{
\item{x}{a bootstrap_light object}

\item{.f}{function to apply to x, can be formula, see the map in the purrr package for details}
}
\description{
creates resampling indices but does not make entire copies of the dataset
}
