\name{YetAnotherRStudioApplication-package}
\alias{YetAnotherRStudioApplication-package}
\alias{YetAnotherRStudioApplication}
\docType{package}
\title{
\packageTitle{YetAnotherRStudioApplication}
}
\description{
\packageDescription{YetAnotherRStudioApplication}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{YetAnotherRStudioApplication}
\packageIndices{YetAnotherRStudioApplication}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{YetAnotherRStudioApplication}

Maintainer: \packageMaintainer{YetAnotherRStudioApplication}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
